refactor(bloc_user): move all user bloc to user folder
