+++ lib/bloc/user/register/register_state.dart
